optimize denoising memory footprint, optimize bge runtime

+++ graxpert/background_extraction.py
@@ -6,34 +6,39 @@ import logging
 from concurrent.futures import wait
 from multiprocessing import shared_memory
 
+import cv2
 import numpy as np
 import onnxruntime as ort
 from astropy.stats import sigma_clipped_stats
 from pykrige.ok import OrdinaryKriging
 from scipy import interpolate, linalg
-from skimage.filters import gaussian
-from skimage.transform import resize
 
+from graxpert.ai_model_handling import get_execution_providers_ordered
 from graxpert.mp_logging import get_logging_queue, worker_configurer
 from graxpert.parallel_processing import executor
 from graxpert.radialbasisinterpolation import RadialBasisInterpolation
-from graxpert.ai_model_handling import get_execution_providers_ordered
+
+
+def gaussian_kernel(sigma=1.0, truncate=4.0):  # follow simulate skimage.filters.gaussian defaults
+    ksize = round(sigma * truncate) - 1 if round(sigma * truncate) % 2 == 0 else round(sigma * truncate)
+    return (ksize, ksize)
 
 
 def extract_background(in_imarray, background_points, interpolation_type, smoothing, downscale_factor, sample_size, RBF_kernel, spline_order, corr_type, ai_path, progress=None):
 
-    shm_imarray = shared_memory.SharedMemory(create=True, size=in_imarray.nbytes)
-    shm_background = shared_memory.SharedMemory(create=True, size=in_imarray.nbytes)
-    imarray = np.ndarray(in_imarray.shape, dtype=np.float32, buffer=shm_imarray.buf)
-    background = np.ndarray(in_imarray.shape, dtype=np.float32, buffer=shm_background.buf)
-    np.copyto(imarray, in_imarray)
+    num_colors = in_imarray.shape[-1]
 
-    num_colors = imarray.shape[-1]
+    shm_imarray = None
+    shm_background = None
 
     if interpolation_type == "AI":
+        imarray = np.ndarray(in_imarray.shape, dtype=np.float32)
+        background = np.ndarray(in_imarray.shape, dtype=np.float32)
+        np.copyto(imarray, in_imarray)
+
         # Shrink and pad to avoid artifacts on borders
         padding = 8
-        imarray_shrink = resize(imarray, output_shape=(256 - 2 * padding, 256 - 2 * padding))
+        imarray_shrink = cv2.resize(imarray, dsize=(256 - 2 * padding, 256 - 2 * padding), interpolation=cv2.INTER_LINEAR)
         imarray_shrink = np.pad(imarray_shrink, ((padding, padding), (padding, padding), (0, 0)), mode="edge")
 
         median = []
@@ -77,7 +82,7 @@ def extract_background(in_imarray, background_points, interpolation_type, smooth
 
         if smoothing != 0:
             sigma = smoothing * 20
-            background = gaussian(image=background, sigma=sigma, channel_axis=-1)
+            background = cv2.GaussianBlur(background, ksize=gaussian_kernel(sigma), sigmaX=sigma, sigmaY=sigma)
 
         if progress is not None:
             progress.update(8)
@@ -96,13 +101,20 @@ def extract_background(in_imarray, background_points, interpolation_type, smooth
         if progress is not None:
             progress.update(8)
 
-        background = gaussian(background, sigma=3.0)  # To simulate tensorflow method='gaussian'
-        background = resize(background, output_shape=(in_imarray.shape[0], in_imarray.shape[1]))
+        sigma = 3.0
+        background = cv2.GaussianBlur(background, ksize=gaussian_kernel(sigma), sigmaX=sigma, sigmaY=sigma)
+        background = cv2.resize(background, dsize=(in_imarray.shape[1], in_imarray.shape[0]), interpolation=cv2.INTER_LINEAR)
 
         if progress is not None:
             progress.update(8)
 
     else:
+        shm_imarray = shared_memory.SharedMemory(create=True, size=in_imarray.nbytes)
+        shm_background = shared_memory.SharedMemory(create=True, size=in_imarray.nbytes)
+        imarray = np.ndarray(in_imarray.shape, dtype=np.float32, buffer=shm_imarray.buf)
+        background = np.ndarray(in_imarray.shape, dtype=np.float32, buffer=shm_background.buf)
+        np.copyto(imarray, in_imarray)
+
         x_sub = np.array(background_points[:, 0], dtype=int)
         y_sub = np.array(background_points[:, 1], dtype=int)
 
@@ -154,15 +166,17 @@ def extract_background(in_imarray, background_points, interpolation_type, smooth
     imarray[:, :, :] = imarray.clip(min=0.0, max=1.0)
 
     in_imarray[:] = imarray[:]
-    background = np.copy(background)
 
     if progress is not None:
         progress.update(8)
 
-    shm_imarray.close()
-    shm_background.close()
-    shm_imarray.unlink()
-    shm_background.unlink()
+    if shm_imarray is not None:
+        shm_imarray.close()
+        shm_imarray.unlink()
+    if shm_background is not None:
+        background = np.copy(background)
+        shm_background.close()
+        shm_background.unlink()
 
     return background
 
@@ -258,7 +272,7 @@ def interpol(shm_imarray_name, shm_background_name, c, x_sub, y_sub, shape, kind
             return
 
         if downscale_factor != 1:
-            result = resize(result, shape, preserve_range=True)
+            result = cv2.resize(src=result, dsize=(shape[1], shape[0]), interpolation=cv2.INTER_LINEAR)
 
         background[:, :, c] = result
     except Exception as e:
